Pull alpine and added user.

+++ cpp/template.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 using namespace atcoder;
 typedef long long ll;
-#define rep(i, n) for (ll i = 0; i < (int)(n); i++)
+#define rep(i, n) for (ll i = 0; i < (ll)(n); i++)
 
 int main()
 {
